+++ webapp/tests/setup/env-setup.js
@@ -3,8 +3,8 @@
  * @tagline         Jest Environment Setup
  * @description     Sets up environment variables for Jest tests
  * @file            webapp/tests/setup/env-setup.js
- * @version         1.6.23
- * @release         2026-02-27
+ * @version         1.6.24
+ * @release         2026-03-06
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/