+++ webapp/tests/unit/i18n/utils/translation-loader.js
@@ -3,8 +3,8 @@
  * @tagline         Load, parse, and flatten translation .conf files to sorted key arrays
  * @description     Utility to load translation files and flatten them to dot-notation key arrays
  * @file            webapp/tests/unit/i18n/utils/translation-loader.js
- * @version         1.6.23
- * @release         2026-02-27
+ * @version         1.6.24
+ * @release         2026-03-06
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/